backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
fastapi-mail==1.4.1
python-multipart==0.0.9
httpx==0.27.0
psycopg[binary]
reportlab==4.4.10